+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -4,7 +4,7 @@ import com.interface21.webmvc.servlet.view.ModelAndView;
 import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import com.interface21.webmvc.servlet.mvc.Controller;
+import com.interface21.webmvc.servlet.controller.Controller;
 
 public class RegisterViewController implements Controller {
