+++ mvc/src/test/java/com/interface21/webmvc/servlet/controller/tobe/AnnotationHandlerMappingTest.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.controller.tobe;
 
 import com.interface21.webmvc.servlet.handler.HandlerExecution;
 import com.interface21.webmvc.servlet.handler.mapping.AnnotationHandlerMapping;
